# api/openapi.yaml
openapi: 3.0.3
info:
  title: Pennywise API
  version: 1.0.0
  description: Self-hosted personal finance dashboard API
servers:
  - url: /api/v1
security:
  - cookieAuth: []
paths:
...
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List transactions with filtering and pagination
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PerPageParam"
        - name: account_id
          in: query
          schema:
            type: string
            format: uuid
        - name: category
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            $ref: "#/components/schemas/TransactionType"
        - name: date_from
          in: query
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          schema:
            type: string
            format: date
        - name: amount_min
          in: query
          schema:
            type: number
            minimum: 0
        - name: amount_max
          in: query
          schema:
            type: number
            minimum: 0
        - name: tags
          in: query
          schema:
            type: string
          description: Comma-separated tag names
        - name: search
          in: query
          schema:
            type: string
        - name: group_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Transaction list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Create a transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionRequest"
      responses:
        "201":
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionResponse"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /transactions/bulk-categorize:
    patch:
      operationId: bulkCategorizeTransactions
      tags: [transactions]
      summary: Update category for multiple transactions at once
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkCategorizeRequest"
      responses:
        "200":
          description: Categories updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkCategorizeResponse"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /transactions/import:
    post:
      operationId: importTransactions
      tags: [transactions]
      summary: Bulk import transactions from CSV
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
                - account_id
              properties:
                file:
                  type: string
                  format: binary
                account_id:
                  type: string
                  format: uuid
      responses:
        "201":
          description: Transactions imported
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResponse"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /transactions/{id}:
    parameters:
      - $ref: "#/components/parameters/IdParam"
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get transaction by ID
      responses:
        "200":
          description: Transaction detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: updateTransaction
      tags: [transactions]
      summary: Update a transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTransactionRequest"
      responses:
        "200":
          description: Transaction updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionResponse"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteTransaction
      tags: [transactions]
      summary: Soft delete a transaction
      responses:
        "204":
          description: Transaction deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List distinct categories used by the current user
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoriesResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
  responses:
    ValidationFailed:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
            - request_id
          properties:
            code:
              $ref: "#/components/schemas/ErrorCode"
            message:
              type: string
            request_id:
              type: string
            details:
              type: object
              additionalProperties: true
    ErrorCode:
      type: string
      enum:
        - VALIDATION_FAILED
        - UNAUTHORIZED
        - FORBIDDEN
        - NOT_FOUND
        - CONFLICT
        - RATE_LIMITED
        - INTERNAL_ERROR
    PaginationMeta:
      type: object
      required:
        - page
        - per_page
        - total
        - total_pages
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
...
    TransactionType:
      type: string
      enum:
        - expense
        - deposit
        - transfer
...
    Currency:
      type: string
      enum:
        - USD
        - BTC
        - ETH
        - EUR
...
    CreateTransactionRequest:
      type: object
      required:
        - type
        - category
        - amount
        - date
        - account_id
      properties:
        type:
          $ref: "#/components/schemas/TransactionType"
        category:
          type: string
          minLength: 1
        tags:
          type: array
          items:
            type: string
        amount:
          type: number
          minimum: 0.01
        currency:
          $ref: "#/components/schemas/Currency"
        date:
          type: string
          format: date
        account_id:
          type: string
          format: uuid
        notes:
          type: string
        is_recurring:
          type: boolean
        group_id:
          type: string
          format: uuid
    UpdateTransactionRequest:
      type: object
      properties:
        type:
          $ref: "#/components/schemas/TransactionType"
        category:
          type: string
          minLength: 1
        tags:
          type: array
          items:
            type: string
        amount:
          type: number
          minimum: 0.01
        currency:
          $ref: "#/components/schemas/Currency"
        date:
          type: string
          format: date
        account_id:
          type: string
          format: uuid
        notes:
          type: string
        is_recurring:
          type: boolean
        group_id:
          type: string
          format: uuid
    TransactionResponse:
      type: object
      required:
        - id
        - user_id
        - account_id
        - type
        - category
        - amount
        - currency
        - date
        - is_recurring
        - tags
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        account_id:
          type: string
          format: uuid
        type:
          $ref: "#/components/schemas/TransactionType"
        category:
          type: string
        amount:
          type: number
        currency:
          $ref: "#/components/schemas/Currency"
        date:
          type: string
          format: date
        notes:
          type: string
        is_recurring:
          type: boolean
        recurring_transaction_id:
          type: string
          format: uuid
        group_id:
          type: string
          format: uuid
        tags:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TransactionListResponse:
      type: object
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/TransactionResponse"
        pagination:
          $ref: "#/components/schemas/PaginationMeta"
    ImportResponse:
      type: object
      required:
        - imported
        - errors
      properties:
        imported:
          type: integer
        errors:
          type: array
          items:
            type: object
            required:
              - row
              - message
            properties:
              row:
                type: integer
              message:
                type: string
...
    BulkCategorizeRequest:
      type: object
      required:
        - updates
      properties:
        updates:
          type: array
          minItems: 1
          maxItems: 200
          items:
            type: object
            required:
              - transaction_id
              - category
            properties:
              transaction_id:
                type: string
                format: uuid
              category:
                type: string
                minLength: 1
    BulkCategorizeResponse:
      type: object
      required:
        - updated
      properties:
        updated:
          type: integer
    CategoriesResponse:
      type: object
      required:
        - categories
      properties:
        categories:
          type: array
          items:
            type: string
...